+++ docker-entrypoint.sh
@@ -12,7 +12,6 @@ then
 fi
 
 python manage.py migrate
-
 # Load default engine types
 python manage.py loaddata fixtures/default_scan_engines.json --app scanEngine.EngineType