+++ startScan/models.py
@@ -3,6 +3,7 @@ from targetApp.models import Domain
 from scanEngine.models import EngineType
 from django.contrib.postgres.fields import JSONField
 
+
 class ScanHistory(models.Model):
     last_scan_date = models.DateTimeField()
     scan_status = models.IntegerField()
@@ -11,7 +12,8 @@ class ScanHistory(models.Model):
 
     def __str__(self):
         # debug purpose remove scan type and id in prod
-        return self.domain_name.domain_name + self.scan_type.scan_type_name + str(self.id)
+        return self.domain_name.domain_name
+
 
 class ScannedHost(models.Model):
     subdomain = models.CharField(max_length=1000)
@@ -29,7 +31,7 @@ class ScannedHost(models.Model):
     takeover = models.CharField(max_length=100, null=True)
 
     def __str__(self):
-        return str(self.subdomain+'---->>>>Scan ID'+str(self.scan_history.id))
+        return str(self.subdomain)
 
 
 class WayBackEndPoint(models.Model):
@@ -41,6 +43,7 @@ class WayBackEndPoint(models.Model):
 
     def __str__(self):
         return self.page_title
+
 
 class ScanActivity(models.Model):
     scan_of = models.ForeignKey(ScanHistory, on_delete=models.CASCADE)