+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
-### V0.2 Release 2020-06-11
+### V0.3 Release 2020-07-21
+
+#### Features
+- YAML based Customization Engine
+- Ability to add wordlists
+- Login Feature
+
+### V0.2 Release 2020-07-11
 
 #### Features
 - Directory Search Enabled
@@ -13,5 +20,5 @@
 - Windows Installation issue fixed
 - Scrollbar Issue on small screens fixed
 
-### V0.1 Release 2020-06-08
+### V0.1 Release 2020-07-08
 - reNgine is released